requirements.txt
Django==5.1.2
python-dotenv==1.0.1
Requests==2.32.3
